requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
yarl==1.6.3
# python-bitbankcc==0.0.4
